Update PWA documentation regarding an issue found in PS 11.9.0
In Platform Server 11.9.0 an issue was found in the stage of the PWA toggle value. It was added as a known issue to the release notes page of that Platform Version page: https://success.outsystems.com/Support/Release_Notes/11/Platform_Server/Platform_Server_11.9.0

When staging an application, the value of the toggle should be staged from the source to the target environment. So if an application has the PWA toggle enabled, when staging that app the target environment should also have the PWA toggle enabled (and vice versa, if the toggle was disabled in the source should be disabled after staging it to the target environment).

A fix for this problem was included in the next Platform Server release, 11.10.0.