#RTAF-3434 - Update version for fix in issue for app feedback not possible to be disabled in Reactive apps
An issue was previously identified for App Feedback in Reactive applications: once App Feedback was enabled, it wasn't possible to disable it. This was documented in the App Feedback main documentation page but now that we have the fix ready for the next version we should update that note so that users know the fix is comming soon.