Edited structure of the content.
It's confusing to the users to have old table records implementation (Traditional only) next to the new Table (Reactive, Mobile only), as they would need to know the differences -- and there's no much benefit, as most new apps are Reactive/Mobile. Therefore, the instructions for older widget are now part of its document in the reference section.

Additionally, we reset the content in the original reference file, as it comes from the source code and can introduce merge conflicts.