Inform users when a light BPT exec is possible (#79)
* Inform users when a light BPT exec is possible

I often find people thinking that once a process is possible to run as a light bpt, all executions will result in a light bpt execution. After confirming in the forums, I state this is not true. To avoid future mistakes, I propose to include this info on this topic. Thanks, <3 ape

Co-authored-by: João Pedro Batista <35229247+JoaoPBatista@users.noreply.github.com>