Update mobile-app-update-scenarios.md
Hello team, the background of this request is related to this ticket #2371698 which was an improvement to be able to determine whether our customers developing mobile apps would need to do automatic or manual generation. These additional notes would help to clarify and our customers to be able to identify when and not manual generation is needed.