Update Mobile App Update Scenarios and SSL Pinning
* Update mobile-app-update-scenarios.md

Hello team, the background of this request is related to this ticket #2371698 which was an improvement to be able to determine whether our customers developing mobile apps would need to do automatic or manual generation. These additional notes would help to clarify and our customers to be able to identify when and not manual generation is needed.

* Content edit

* Tag fix

* Added info related to external configuration

By design, the platform ignores changes in the external configuration files of non-core components as a trigger to generate a new build. This update clarifies that in the Mobile App Update Scenarios document and adds two reminders in the SSL Pinning plugin document.

Co-authored-by: Romeo Mlinar <30351511+rmlos@users.noreply.github.com>
Co-authored-by: Romeo Mlinar <romeo.mlinar@outsystems.com>