Removed a sentence
The documentation team has yet to receive more drafts from the field to be able to extract common logic and add examples to this document. Still, all critical information is outlined in "How to develop an app that updates only through the app stores". We're removing the sentence "OutSystems is preparing more documentation to guide you in creating apps for resilient updates." as it introduces the doubt whether the document is complete.