Add a new section on Background Job
I included a new section named "Create logic to define a background job" that describes how to use the newly added SetBackgroundJob action. The description includes examples of two use cases with corresponding images of the correct parametrization. 

This is to be updated when the new version of the plugin with these functionalities becomes available on Forge. Please check that with Gonçalo Saraiva (goncalo.saraiva@outsystems.com).