Add a new section on Background Job (#232)
* Add a new section on Background Job

I included a new section named "Create logic to define a background job" that describes how to use the newly added SetBackgroundJob action. The description includes examples of two use cases with corresponding images of the correct parametrization. 

This is to be updated when the new version of the plugin with these functionalities becomes available on Forge. Please check that with Gonçalo Saraiva (goncalo.saraiva@outsystems.com).

* New version of the Background job section

Edit the Background job section following the feedback provided by Jimmy Goldfarb.

Thank you Jimmy for your suggestions!

* Solving the issue of images breaking the text

Add a newline before the images