Update upgrade-platform-module-prep.md
I didn't get enough clarity in this article that the module dependencies are being upgraded to the latest version of "system components". We talk about "platform version" in the first paragraphs but this is not very precise. It needs to be more clear or said in a different way that modules continue to run under the old cached DLLs until you decide to deploy them. What the Modules Preparation is doing is updating not just the platform version/system component dependencies but ALL dependencies and compiling the new DLLs for them. The fact that all dependencies get updated is why the preparation might fail; for example, if a strong dependency is broken or a developer is in the middle of work in a producer module.

I hope my changes will improve this understanding.