OnReady before Render?
Confirm if I was wrong but, in the phrase "The On Ready event handler runs when the Screen or Block is ready, i.e. when the DOM is ready, after the first render." 
I think the last part is wrong "after the first render". Wouldn't it be: "before the first render"?