Clarified Add System Dependencies permissions
The documentation was unclear that someone without Add System Dependencies permissions could use the dependencies after they are added to the modules. We aim to clarify that you need permission to add the dependencies, but after they are added anyone who can edit the module can use the added dependencies.