Update responsive.md
Currently, the classes that refer to Browser and Touch Detection are only applied to Reactive Web apps. We have created RPM-5340 to see if this is the intended behaviour or not. In the meantime, this documentation should be updated to avoid misleading customers.  If this is the desired behaviour, then the changes in the documentation should be permanent.